to editable category contents

+++ database/migrations/2023_09_02_105455_create_record_states_table.php
@@ -19,7 +19,8 @@ return new class extends Migration
             // cascadeOnDelete()は親のテーブルのレコードが削除された場合に一緒に削除
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             //外部キーの制約（null許容）はforeignIdの直後
-            $table->integer('bodyWeight')->nullable();
+            //小数点あり：float('カラム名', 小数点を含めた桁数, 小数点以下の桁数)
+            $table->float('bodyWeight',8, 1)->nullable();
             $table->date('recorded_at');
             $table->timestamps();                        
         });
